CI: Pass npm.tgz as a relative path [skip travis]

+++ node/scripts/test-npm.sh
@@ -4,6 +4,10 @@ scriptDir="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
 echo $scriptDir$
 
 npmTGZFile=$1
+if [ ! -f "$npmTGZFile" ]; then
+    echo "The npmTGZFile file does not exist: $npmTGZFile"
+    exit 1
+fi
 
 rm -rf ./npm_test && cp -r $scriptDir/../npm/test ./npm_test
 ls npm_test
